+++ web-api-server-rx-module/src/META-INF/rx-module.properties
@@ -1,1 +1,1 @@
-wire-module=hiconic.rx.web.undertow.wire.WebApiServerRxModule
+wire-module=hiconic.rx.web.ddra.wire.WebApiServerRxModule
