+++ web-api-server-rx-module/src/hiconic/rx/web/ddra/wire/WebApiServerRxModule.java
@@ -1,7 +1,7 @@
-package hiconic.rx.web.undertow.wire;
+package hiconic.rx.web.ddra.wire;
 
 import hiconic.rx.module.api.wire.RxModule;
-import hiconic.rx.web.undertow.wire.space.WebApiServerRxModuleSpace;
+import hiconic.rx.web.ddra.wire.space.WebApiServerRxModuleSpace;
 
 public enum WebApiServerRxModule implements RxModule<WebApiServerRxModuleSpace> {
 	INSTANCE;
